Adjust file structure for tree code
Also make error handling better for inclusion proof generation/verification
